+++ src/adapter/http/src/data/router.rs
@@ -17,13 +17,16 @@
 // by the Apache License, Version 2.0.
 
 use super::ingest_handler::dataset_ingest_handler;
-use super::query_handler::dataset_query_handler;
+use super::query_handler::{dataset_query_handler, dataset_query_handler_post};
 use super::tail_handler::dataset_tail_handler;
 
 /////////////////////////////////////////////////////////////////////////////////
 
 pub fn root_router() -> axum::Router {
-    axum::Router::new().route("/query", axum::routing::get(dataset_query_handler))
+    axum::Router::new().route(
+        "/query",
+        axum::routing::get(dataset_query_handler).post(dataset_query_handler_post),
+    )
 }
 
 /////////////////////////////////////////////////////////////////////////////////